refactore: serializers to rdm-records structure
* to align the directory structure of resources/serializers to the
  directory structure of invenio-rdm-records.

* in contrast to invenio-rdm-records the serializer classes are moved to
  a separate serializer.py file so that __init__.py does not contain
  actual code.
