+++ invenio_records_lom/resources/config.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2021 Graz University of Technology.
+# Copyright (C) 2021-2024 Graz University of Technology.
 #
 # invenio-records-lom is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -13,11 +13,11 @@ from invenio_rdm_records.resources import IIIFResourceConfig
 from invenio_records_resources.resources import RecordResourceConfig
 from invenio_records_resources.resources.files import FileResourceConfig
 
-from .serializers import LOMUIJSONSerializer
+from .serializers import LOMToUIJSONSerializer
 
 record_serializer = {
     "application/json": ResponseHandler(JSONSerializer()),
-    "application/vnd.inveniolom.v1+json": ResponseHandler(LOMUIJSONSerializer()),
+    "application/vnd.inveniolom.v1+json": ResponseHandler(LOMToUIJSONSerializer()),
 }
 
 url_prefix = "/oer"
