+++ invenio_records_lom/resources/serializers/dublincore/schema.py
