+++ invenio_records_lom/resources/serializers/oai/schema.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2023 Graz University of Technology.
+# Copyright (C) 2023-2024 Graz University of Technology.
 #
 # invenio-records-lom is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -334,7 +334,7 @@ class ClassificationSchema(ExcludeUnknownOrderedSchema):
 #
 # main schema
 #
-class LOMMetadataToOAISchema(ExcludeUnknownOrderedSchema):
+class LOMToOAISchema(ExcludeUnknownOrderedSchema):
     """Schema to reorder fields and exclude unneeded fields."""
 
     general = fields.Nested(GeneralSchema(), required=True)
