+++ invenio_records_lom/resources/serializers/ui/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2024 Graz University of Technology.
+#
+# invenio-records-lom is free software; you can redistribute it and/or modify it
+# under the terms of the MIT License; see LICENSE file for more details.
+
+"""UI Serializer."""
+
+from .serializer import LOMToUIJSONSerializer
+
+__all__ = ("LOMToUIJSONSerializer",)
