+++ invenio_records_lom/resources/serializers/ui/schema.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2021-2023 Graz University of Technology.
+# Copyright (C) 2021-2024 Graz University of Technology.
 #
 # invenio-records-lom is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -24,7 +24,7 @@ from marshmallow_oneofschema import OneOfSchema
 from werkzeug.local import LocalProxy
 
 from ....services.schemas.fields import ControlledVocabularyField
-from .utils import get_newest_part, get_related, get_text
+from ..utils import get_newest_part, get_related, get_text
 
 
 class Title(fields.Field):
